requirements.txt
azure-ai-ml>=1.23.0
azure-identity
azure-mgmt-resource
azure-mgmt-sql
pyodbc
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
mlflow>=2.12.0
azureml-mlflow
matplotlib
seaborn
requests
flake8
pytest
pytest-cov
